feat: use dependency injection with callback function
The database class should be implementation-agnostic, so the execute function needs to support arbitrary callbacks. There are currently two versions of execute in the event the user does not care about the result (e.g. when inserting), but a default value in the templated version is a sound alternative.

The templated version is defined within the header file for C++ compiler reasons. The alternative is to explcitly define each possible ttype of T in the source file; right now, the only value is std::string, but this doesn't seem idiomatic.